set beneficiary, set records
