+++ contracts/subdomainregistrar/SubdomainRegistrar.sol
@@ -3,17 +3,23 @@ pragma solidity ^0.8.4;
 
 import "../wrapper/INameWrapper.sol";
 import "@openzeppelin/contracts/utils/Address.sol";
+import { ERC1155Holder } from "@openzeppelin/contracts/token/ERC1155/utils/ERC1155Holder.sol";
+import "hardhat/console.sol";
 
 error Unavailable();
 error Unauthorised(bytes32 node);
 error NotEnoughEther();
+error NameNotRegistered();
 
-contract SubdomainRegistrar {
+contract SubdomainRegistrar is ERC1155Holder {
     INameWrapper public immutable wrapper;
     using Address for address;
 
     mapping(bytes32 => uint256) public registrationFees;
+    mapping(bytes32 => address) public beneficiaries;
     mapping(bytes32 => uint256) public expiries;
+
+    event NameRenewed(bytes32 node, uint256 duration);
 
     constructor(INameWrapper _wrapper) {
         wrapper = _wrapper;
@@ -26,12 +32,16 @@ contract SubdomainRegistrar {
         _;
     }
 
+    function setupDomain(bytes32 node, uint256 fee, address beneficiary) public onlyOwner(node) {
+        setRegistrationFee(node, fee);
+        beneficiaries[node] = beneficiary;
+    }
+
     function setRegistrationFee(bytes32 node, uint256 fee)
         public
         onlyOwner(node)
     {
         registrationFees[node] = fee;
-        // set address to receive fees
     }
 
     function available(bytes32 node) public view returns (bool) {
@@ -51,12 +61,18 @@ contract SubdomainRegistrar {
     ) public payable {
         bytes32 labelhash = keccak256(bytes(label));
         bytes32 node = keccak256(abi.encodePacked(parentNode, labelhash));
-        uint256 registrationFee = duration * registrationFees[node];
+        uint256 registrationFee = duration * registrationFees[parentNode];
+
         if (!available(node)) {
             revert Unavailable();
         }
         if (msg.value < registrationFee) {
             revert NotEnoughEther();
+        }
+
+        if(records.length > 0){
+            wrapper.setSubnodeOwner(parentNode, label, address(this), 0);
+            _setRecords(node, resolver, records);
         }
 
         wrapper.setSubnodeRecord(
@@ -67,17 +83,25 @@ contract SubdomainRegistrar {
             ttl,
             _fuses | PARENT_CANNOT_CONTROL // burn the ability for the parent to control
         );
-
-        _setRecords(node, resolver, records);
+        
         expiries[node] = block.timestamp + duration;
 
-        // Transfer fees - could allow the fees to go elsewhere
-        address owner = wrapper.ownerOf(uint256(node));
-        (bool sent, ) = owner.call{value: registrationFee}("");
+        (bool sent, ) = beneficiaries[parentNode].call{value: registrationFee}("");
 
         if (!sent) {
             revert();
         }
+    }
+
+    function renew(bytes32 id, uint duration) external returns(uint) {
+        if(expiries[id] < block.timestamp) {
+            revert NameNotRegistered();
+        }
+        require(expiries[id] + duration > duration); // Prevent future overflow
+
+        expiries[id] += duration;
+        emit NameRenewed(id, expiries[id]);
+        return expiries[id];
     }
 
     function _setRecords(
